Fix leaks when polling for PR mergeability
The GitHub API requires polling PR status if the mergeable property is
null, because a background job is running to compute this value. But the
mergeable property is also null for merged PRs. In cases where there was
a race between multiple requests to merge the PR (e.g. if there were
many status updates at once), one or more of these requests would end up
polling a merged PR forever, waiting for it to be mergeable.

Add a limit on the total number of poll attempts and also check for PRs
that are merged or closed after we start polling. In addition, only try
to merge PRs in response to successful status checks.