Add support for multiple config file paths
Based on server configuration, the application will now look for the
repository configuration file in multiple paths until it finds one. If
no paths are specified, the existing .bulldozer.yml is used.

This commit makes the minimal amount of change necessary to get this
working. Ideally, there would be more refactoring to standardize config
passing and client construction, but I'm not ready to do that yet.