Evaluate PR Body independently from comments
Right now we compare comment substrings on both comments and the PR body. In some cases,
it should be possible to lock down evaluation to only the PR body, instead of comments
_and_ the pr body.

Users who submit PRs do not have write access to the repo, so they cannot add a label.
For comments, anyone with read access to the repo can add a comment on any PR.

However, only someone with write access (or the Author) can edit the PR description.