Support strategies for generating commit titles (#101)
When squashing commits, users can now set how the commit title is
generated, since GitHub creates different titles based on the number of
commits to merge. The "pull_request_title" strategy uses the title of
the pull request, the "first_commit_title" strategy uses the first line
of the first commit in the pull request, and the "github_default_title"
strategy defers to GitHub's logic. The title always end with the PR
number in parentheses.

Note that this changes the default behavior for generating titles from
"github_default_title" to "pull_request_title". I think this new
default is more user-friendly (and expected), but it is different.

This also does some refactoring, moving commit listing to pull.Context
and clarifying that merge options only apply to squashes.