Add option to prevent merges without required checks (#217)
This PR includes changes which seek to resolve #122, where PRs without any required checks are merged automatically, often due to mis-configuration.

A new option is added to the v1 configuration format to allow/deny merges with no required status checks. The option is nested within the `merge` object:
```yaml
version: 1

merge:
  allow_merge_with_no_checks: false
```

The v0 configuration format is not updated to support this feature. Repositories using the v0 configuration will be treated as if `allow_merge_with_no_checks` is set to the default value `false`.