Set HTTP config values from env vars
Heroku expects the web process to dynamically set the listen port based
on the PORT variable. This, in combination with
a [.profile](https://devcenter.heroku.com/articles/dynos#the-profile-file)
to set `PORT=BULLDOZER_PORT`, will make it easier to run bulldozer in
Heroku, as well as allowing other HTTP-related config to be set via
environment variables.