Set HTTP config values from env vars (#220)
Heroku expects the web process to dynamically set the listen port based
on the PORT variable. This, in combination with a .profile file to set
`PORT=BULLDOZER_PORT`, will make it easier to run Bulldozer in Heroku,
as well as allowing other HTTP-related config to be set via environment
variables.