Update PR when labeled/opened
This will conditionally update a PR whenever the PR is labeled or opened, which
will partially fulfill [this
suggestion](https://github.com/palantir/bulldozer/issues/145#issuecomment-520662592)
for https://github.com/palantir/bulldozer/issues/145.

It doesn't seem possible to determine which label was added via the
Github webhook payload, so this will update the PR regardless of whether
the label added is what is configured in `bulldozer.yml`.