Fix prematurely merging PR after update
After my changes in https://github.com/palantir/bulldozer/pull/235,
I noticed that sometimes a PR is updated and then merged as part of the
same webhook request, before the required statuses are fulfilled. This
is because bulldozer was still using the stale statuses from before the
update was performed.