Fix processing for check run events (#250)
In 1.13.0, an unrelated refactor used an incomplete pull request object
to fetch configuration, resulting in requests that are missing the
organization name. Because the config isn't loaded, all further
processing for check run events is skipped.

Use the complete PR object and also re-order things to make it less
tempting to use the plain PR object once the full one is available.