Implement remote and default config loading
Use the go-githubapp/appconfig package to support remote configuration
references and organization-level defaults. To make this simpler to
implement, there are two changes to existing behavior:

1. If any config is invalid, Bulldozer stops processing the PR and logs
   a warning. Previously, it would fall back to global config (if set)
   in this case.

2. All defined config paths may contain either v1 or v0 configuration.
   Previously, some paths could only contain one version or the other.
   Since v0 config is undocumented and has been deprecated for a
   long-time, I don't expect this to impact anything.

I also did some refactoring, moving the fetching logic to the handler
package while leaving the parsing logic in the bulldozer package. This
mirrors the layout in policy-bot.